[state-sync] moving client to state_sync_client.rs
