+++ state-synchronizer/src/synchronizer.rs
@@ -1,37 +1,29 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
+
 use crate::{
-    coordinator::{CoordinatorMessage, SyncCoordinator, SyncRequest},
+    coordinator::SyncCoordinator,
     executor_proxy::{ExecutorProxy, ExecutorProxyTrait},
     network::{StateSynchronizerEvents, StateSynchronizerSender},
-    SynchronizerState,
+    state_sync_client::CoordinatorMessage,
+    state_sync_client::StateSyncClient,
 };
-use anyhow::{format_err, Result};
+use anyhow::Result;
 use executor_types::ChunkExecutor;
 use futures::{
     channel::{mpsc, oneshot},
-    future::Future,
     SinkExt,
 };
 use libra_config::config::{NodeConfig, RoleType, StateSyncConfig, UpstreamConfig};
-use libra_mempool::{CommitNotification, CommitResponse};
-use libra_types::{
-    contract_event::ContractEvent, ledger_info::LedgerInfoWithSignatures, transaction::Transaction,
-    waypoint::Waypoint, PeerId,
-};
-use std::{
-    boxed::Box,
-    collections::HashMap,
-    sync::Arc,
-    time::{Duration, SystemTime},
-};
+use libra_mempool::CommitNotification;
+use libra_types::{waypoint::Waypoint, PeerId};
+use std::{boxed::Box, collections::HashMap, sync::Arc};
 use storage_interface::DbReader;
 use subscription_service::ReconfigSubscription;
-use tokio::{
-    runtime::{Builder, Runtime},
-    time::timeout,
-};
+use tokio::runtime::{Builder, Runtime};
 
+/// StateSynchronizer is in charge of taking requests from network clients (other full nodes),
+/// as well as internal clients (consensus)
 pub struct StateSynchronizer {
     _runtime: Runtime,
     coordinator_sender: mpsc::UnboundedSender<CoordinatorMessage>,
@@ -109,9 +101,7 @@ impl StateSynchronizer {
     }
 
     pub fn create_client(&self) -> Arc<StateSyncClient> {
-        Arc::new(StateSyncClient {
-            coordinator_sender: self.coordinator_sender.clone(),
-        })
+        Arc::new(StateSyncClient::new(self.coordinator_sender.clone()))
     }
 
     /// The function returns a future that is fulfilled when the state synchronizer is
@@ -123,78 +113,5 @@ impl StateSynchronizer {
             .send(CoordinatorMessage::WaitInitialize(cb_sender))
             .await?;
         cb_receiver.await?
-    }
-}
-
-pub struct StateSyncClient {
-    coordinator_sender: mpsc::UnboundedSender<CoordinatorMessage>,
-}
-
-impl StateSyncClient {
-    /// Sync validator's state to target.
-    /// In case of success (`Result::Ok`) the LI of storage is at the given target.
-    /// In case of failure (`Result::Error`) the LI of storage remains unchanged, and the validator
-    /// can assume there were no modifications to the storage made.
-    /// It is up to state synchronizer to decide about the specific criteria for the failure
-    /// (e.g., lack of progress with all of the peer validators).
-    pub fn sync_to(&self, target: LedgerInfoWithSignatures) -> impl Future<Output = Result<()>> {
-        let mut sender = self.coordinator_sender.clone();
-        let (callback, cb_receiver) = oneshot::channel();
-        let request = SyncRequest {
-            callback,
-            target,
-            last_progress_tst: SystemTime::now(),
-        };
-        async move {
-            sender
-                .send(CoordinatorMessage::Request(Box::new(request)))
-                .await?;
-            cb_receiver.await?
-        }
-    }
-
-    /// Notifies state synchronizer about new version
-    pub fn commit(
-        &self,
-        // *successfully* committed transactions
-        committed_txns: Vec<Transaction>,
-        reconfig_events: Vec<ContractEvent>,
-    ) -> impl Future<Output = Result<()>> {
-        let mut sender = self.coordinator_sender.clone();
-        async move {
-            let (callback, callback_rcv) = oneshot::channel();
-            sender
-                .send(CoordinatorMessage::Commit(
-                    committed_txns,
-                    reconfig_events,
-                    callback,
-                ))
-                .await?;
-
-            match timeout(Duration::from_secs(5), callback_rcv).await {
-                Err(_) => {
-                    Err(format_err!("[state sync client] failed to receive commit ACK from state synchronizer on time"))
-                }
-                Ok(resp) => {
-                    let CommitResponse { msg } = resp??;
-                    if msg != "" {
-                        Err(format_err!("[state sync client] commit failed: {:?}", msg))
-                    } else {
-                        Ok(())
-                    }
-                }
-            }
-        }
-    }
-
-    /// Returns information about StateSynchronizer internal state
-    pub fn get_state(&self) -> impl Future<Output = Result<SynchronizerState>> {
-        let mut sender = self.coordinator_sender.clone();
-        let (cb_sender, cb_receiver) = oneshot::channel();
-        async move {
-            sender.send(CoordinatorMessage::GetState(cb_sender)).await?;
-            let info = cb_receiver.await?;
-            Ok(info)
-        }
     }
 }
